save predictions and loss on train and validation set for a model trained to 92.6 accuracy
