+++ fastai/imagito/good_model/metrics.csv
@@ -1,0 +1,21 @@
+epoch,train_loss,valid_loss,accuracy,time
+0,1.319795,1.363368,0.548000,05:57
+1,1.151856,1.034766,0.672000,05:36
+2,1.095174,1.518175,0.538000,05:36
+3,0.979072,1.155515,0.660000,05:36
+4,0.927019,1.184804,0.658000,05:37
+5,0.812842,0.983508,0.716000,05:36
+6,0.730610,0.874821,0.738000,05:37
+7,0.676148,0.853182,0.734000,05:35
+8,0.583525,0.623572,0.822000,05:36
+9,0.532756,0.686327,0.784000,05:35
+10,0.512661,0.379943,0.888000,05:35
+11,0.447488,0.527481,0.840000,05:36
+12,0.403702,0.358054,0.890000,05:35
+13,0.338214,0.329368,0.904000,05:36
+14,0.289028,0.289420,0.902000,05:35
+15,0.268885,0.304213,0.902000,05:34
+16,0.231736,0.265661,0.918000,05:35
+17,0.213361,0.239422,0.930000,05:35
+18,0.219712,0.246185,0.926000,05:35
+19,0.186830,0.243573,0.926000,05:36
